% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RPackage.R
\name{citeIneq}
\alias{citeIneq}
\title{citeIneq}
\usage{
citeIneq(result, metric = "all", type = "resampled", showMargins = FALSE)
}
\arguments{
\item{result}{a \code{adjustCiteMetrics} class object created by \code{adjustCiteMetrics}.}

\item{metric}{character vector indicating the inequality measures to report; possible choices are \verb{everCited', }gini', \verb{hhi', or user-requested quantiles, such as }q20' or \verb{q80'.  Default is }all', which reports all metrics computed by \code{adjustCiteMetrics}.}

\item{type}{character string, the default, \verb{resampled', reports the adjusted inequality measures.  Set to }uncorrected' to see the inequality metrics before sampling adjustment is applied}

\item{showMargins}{logical, report information of the marginal and reference number of papers and citations (default is FALSE)}
}
\description{
Extract citation inequality metrics from an \code{adjustCiteMetrics} object after resampling to correct for marginals bias
}
\examples{
citeIneq(result) 
}
