% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RPackage.R
\name{adjustCiteMetrics}
\alias{adjustCiteMetrics}
\title{adjustCiteMetrics}
\usage{
adjustCiteMetrics(
  papers,
  cites,
  pubID = "paperID",
  pubYear = "publishedYear",
  citedID = "citedPaperID",
  citedYear = "citedPaperYear",
  citingYear = NULL,
  citationWindow = NULL,
  quantiles = c(0.2, 0.8),
  refYear = NULL,
  refPaperCount = NULL,
  refCiteCount = NULL,
  sims = 10,
  lowCitesInclusion = FALSE,
  lowCitesThreshold = 0.1,
  paperDupThreshold = 0.95,
  periods = NULL,
  verbose = FALSE
)
}
\arguments{
\item{papers}{required dataframe containing all citable papers studied (one row per paper). This dataframe must include two columns: the published paper IDs (\code{paperID}) and the publication year of the paper (\code{publishedYear}).}

\item{cites}{required dataframe containing all citations made to the papers listed in \code{papers} over a given set of years (one row per citation). This dataframe must include at least two columns: the IDs of the cited papers (\code{citedPaperID}), corresponding to the IDs given in \code{papers}, and the cited paper's publication year (\code{citedPaperYear}). Optionally, this dataframe may include the published year of paper sending the citation (\code{citingYear}) may be included optional.  \code{citingYear} must be included if users wish to filter citations made within \code{citationWindow} years of the cited papers' publication dates.}

\item{pubID}{character string, the column of \code{papers} containing publication IDs (default is `paperID').  IDs may be numeric, character, or factor.}

\item{pubYear}{character string, the column of \code{papers} containing publication years of each paper in numeric format (default is `publishedYear')}

\item{citedID}{character string, the column of \code{cites} containing the IDs of cited papers (default is `citedPaperID'). IDs may be numeric, character, or factor.}

\item{citedYear}{character string, the column of \code{cites} containing the years (as numeric data) in which papers were cited (default is `citedPaperYear')}

\item{citingYear}{character string, the column of \code{cites} containing the year (as numeric data) in which each citing paper was published (default = NULL)}

\item{citationWindow}{numeric scalar, the number of years covered by the citation window. For example, 2 means the function only analyzes citations made within 2 years after the cited paper is published. (default = NULL)}

\item{quantiles}{numeric vector in the interval \(0,1] indicating the citation quantiles to be computed. Default = c(.20, .80), for the percentage of papers accounting for 20 percent and 80 percent of all citations.}

\item{refYear}{numeric scalar, the reference (or baseline) year among available publication years of papers in \code{papers}. Either \code{refYear} or \code{refPaperCount}/\code{refCiteCount} must be provided. (default = NULL)}

\item{refPaperCount}{a numeric scalar identifying the referenced publication count. Either \code{refYear} or \code{refPaperCount}/\code{refCiteCount} must be provided. (default = NULL)}

\item{refCiteCount}{a numeric scalar identifying the referenced citation count. Either \code{refYear} or \code{refPaperCount}/\code{refCiteCount} must be provided. (default = NULL)}

\item{sims}{optional; a numeric scalar identifying the number of times to run the simulation (default = 10)}

\item{lowCitesInclusion}{logical, TRUE: reports all results; FALSE: excludes any years from the results where the total number of citations is lower than the reference year in at least \code{lowCitesThreshold} percent of simulations (default=FALSE)}

\item{lowCitesThreshold}{numeric scalar in the interval [0,1] identifying threshold to define lowCites (default = 0.1)}

\item{paperDupThreshold}{numeric scalar in the interval \(0,1] (default = 0.95)}

\item{periods}{optional; numeric vectors identifying which years we analyze in the data. If it is NULL (default), the function automatically selects all years in \code{papers}.}

\item{verbose}{logical; returns detailed warnings related to lowCites (default=FALSE)}
}
\description{
Use resampling to correct citation inequality measures for marginals bias
}
\section{Details}{

Are these details?
}

\section{Values}{

Are these values?
}

\section{References}{

Are these references?
}

\section{See also}{

Are these see also?
}

\examples{
data(papers)
data(cites)
result <- adjustCiteMetrics(papers = papers, 
cites = cites, 
pubID = "paperID", 
pubYear = "publishedYear", 
citedID = "citedID",
citedYear = "citedYear",
refYear = 1996)
}
