% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RPackage.R
\name{adjustCiteMetrics}
\alias{adjustCiteMetrics}
\title{Adjust function}
\usage{
adjustCiteMetrics(
  papers,
  cites,
  pubID = "paperID",
  pubYear = "publishedYear",
  citedID = "citedPaperID",
  citedYear = "citedPaperYear",
  citingYear = NULL,
  citationWindow = NULL,
  quantiles = c(0.2, 0.8),
  refYear = NULL,
  refPaperCount = NULL,
  refCiteCount = NULL,
  sims = 10,
  lowCitesInclusion = FALSE,
  lowCitesThreshold = 0.1,
  paperDupThreshold = 0.95,
  periods = NULL,
  verbose = FALSE
)
}
\description{
Adjust function
}
