% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RPackage.R
\name{citeIneq}
\alias{citeIneq}
\title{citeIneq}
\usage{
citeIneq(result, metric = "all", type = "resampled", showMargins = FALSE)
}
\arguments{
\item{result}{\code{adjustCiteMetrics} class. It should be created by \code{adjustCiteMetrics} function.}

\item{metric}{optional; character string to choose which inequality measures to report. default option ("all") reports all metrics computed by \code{adjustCiteMetrics}.}

\item{type}{optional; character string to decide which metrics to show between resampled (=adjusted) and original measures (=uncorrected). (default = "resampled")}

\item{showMargins}{optional; logical value to decide whether to show margins or not.}
}
\description{
Extract
}
